refactor: split napi mdast raw serialization
